ua-sources/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
aiohttp==3.9.3
pyyaml==6.0.1
requests==2.31.0
python-multipart==0.0.9
email-validator==2.1.0
jinja2==3.1.3

# Data Pipeline (Phase 3)
minio==7.2.3
pandas==2.1.4
openpyxl==3.1.2
asyncpg==0.29.0
opensearch-py==2.4.2

# Security & Auth (Phase 5)
PyJWT[crypto]==2.8.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# Semantic Search (Phase 6-7)
sentence-transformers>=3.0.0
qdrant-client==1.7.0
numpy==1.24.3
Pillow>=10.0.0

# ML/AI (Optional for local)
transformers==4.37.0
torch==2.1.2

# Task Queue
celery==5.3.6
redis==5.0.1

# Database & Async Utils
greenlet>=3.0.3
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.1
psutil==5.9.8
httpx>=0.25.0

# Monitoring
prometheus-client>=0.19.0

# ═══════════════════════════════════════════════════════════════
# НОВI ЗАЛЕЖНОСТI ДЛЯ ТЗ ІНТЕГРАЦІЇ (2025-12-06)
# ═══════════════════════════════════════════════════════════════

# ML/NLP Advanced
spacy>=3.7.0               # NER та linguistic features
https://github.com/explosion/spacy-models/releases/download/uk_core_news_sm-3.7.0/uk_core_news_sm-3.7.0-py3-none-any.whl
nlpaug>=1.1.11             # Data augmentation
sentencepiece>=0.1.99      # Tokenizer for T5/BART

# Integrations
slack-sdk>=3.27.0          # Slack API client
notion-client>=2.0.0
google-api-python-client>=2.118.0
kubernetes>=29.0.0
prometheus-api-client>=0.5.5
google-auth>=2.25.0        # Google OAuth2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Security Enhancements
argon2-cffi>=23.1.0        # Password hashing (stronger than bcrypt)
cryptography>=41.0.0       # Token encryption

# Observability
jaeger-client>=4.8.0       # Distributed tracing (optional)
opentelemetry-api>=1.21.0  # OpenTelemetry support

# Data Processing
python-docx>=1.1.0         # DOCX parsing for Drive import
pdfplumber>=0.10.0         # PDF parsing
beautifulsoup4>=4.12.3     # HTML parsing
lxml>=5.1.0                # Fast XML/HTML parser

# MLOps & Self-Improvement (v22.0)
dvc[s3]>=3.40.0            # Data Version Control
augly>=1.0.0               # Facebook AugLy (Optional)
mlflow>=2.10.0             # Experiment Tracking
